requirements.txt
matplotlib==3.7.1
mmcv==1.7.1
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
scikit_image==0.19.1
timm==0.6.12
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
